test(liveliness): split candidate-a repro into per-test binaries
Each candidate_a test deliberately leaks its session/tokens/thread to
keep query-hang detection uncontaminated by teardown cost. Running all
four as #[test]s in one process meant leaked state from earlier tests
could bleed into later ones -- confirmed directly: the router-mode
oversized-handler control hung when run after 3 other tests in the
same process, but passed cleanly (0.53s) in isolation.

Cargo treats each direct tests/*.rs file as its own binary/process, so
splitting into four single-test files gives real process isolation
without needing nextest or manual subprocess forking. Shared setup
lives in tests/liveliness_self_deadlock_support/mod.rs, which cargo's
default tests/*.rs discovery does not pick up as its own binary.
